CI will fail with skipped exams, closes #55

+++ spec/rails_helper.rb
@@ -46,7 +46,11 @@ RSpec.configure do |config|
   config.include ActiveSupport::Testing::TimeHelpers
   config.include Warden::Test::Helpers
 
-  config.before(focus: true) { fail 'Remove focused specs before commit!' if ENV['CI'] }
+  if ENV['CI']
+    config.before(focus: true) { fail 'Remove focused specs before commit!' }
+    config.before(skip: true) { fail 'Remove skipped specs before commit!' }
+  end
+
   config.after(:each) { Warden.test_reset! }
   config.after(:suite) { FileUtils.rm_rf("#{Rails.root}/tmp/paperclip") }
